+++ src/app/api/kamu.graphql.interface.ts
@@ -126,6 +126,11 @@ export type AuthMutLoginArgs = {
     loginMethod: Scalars["String"];
 };
 
+export type BatchingConditionInput = {
+    maxBatchingInterval: TimeDeltaInput;
+    minRecordsToAwait: Scalars["Int"];
+};
+
 export type BlockRef = {
     __typename?: "BlockRef";
     blockHash: Scalars["Multihash"];
@@ -391,10 +396,9 @@ export type DatasetFlowConfigsMutResumeFlowsArgs = {
 };
 
 export type DatasetFlowConfigsMutSetConfigBatchingArgs = {
+    batching: BatchingConditionInput;
     datasetFlowType: DatasetFlowType;
-    minimalDataBatch?: InputMaybe<Scalars["Int"]>;
     paused: Scalars["Boolean"];
-    throttlingPeriod?: InputMaybe<TimeDeltaInput>;
 };
 
 export type DatasetFlowConfigsMutSetConfigScheduleArgs = {
@@ -748,8 +752,8 @@ export type FlowConfiguration = {
 
 export type FlowConfigurationBatching = {
     __typename?: "FlowConfigurationBatching";
-    minimalDataBatch?: Maybe<Scalars["Int"]>;
-    throttlingPeriod?: Maybe<TimeDelta>;
+    maxBatchingInterval: TimeDelta;
+    minRecordsToAwait: Scalars["Int"];
 };
 
 export type FlowConfigurationSchedule = Cron5ComponentExpression | TimeDelta;
@@ -808,6 +812,7 @@ export type FlowDescriptionUpdateResult = {
     __typename?: "FlowDescriptionUpdateResult";
     numBlocks: Scalars["Int"];
     numRecords: Scalars["Int"];
+    updatedWatermark?: Maybe<Scalars["DateTime"]>;
 };
 
 export type FlowEdge = {
@@ -840,11 +845,11 @@ export type FlowEventQueued = FlowEvent & {
     eventTime: Scalars["DateTime"];
 };
 
-export type FlowEventStartConditionDefined = FlowEvent & {
-    __typename?: "FlowEventStartConditionDefined";
+export type FlowEventStartConditionUpdated = FlowEvent & {
+    __typename?: "FlowEventStartConditionUpdated";
     eventId: Scalars["EventID"];
     eventTime: Scalars["DateTime"];
-    startCondition: FlowStartCondition;
+    startConditionKind?: Maybe<FlowStartConditionKind>;
 };
 
 export type FlowEventTaskChanged = FlowEvent & {
@@ -871,6 +876,12 @@ export type FlowIncompatibleDatasetKind = SetFlowConfigResult &
         message: Scalars["String"];
     };
 
+export type FlowInvalidBatchingConfig = SetFlowConfigResult & {
+    __typename?: "FlowInvalidBatchingConfig";
+    message: Scalars["String"];
+    reason: Scalars["String"];
+};
+
 export type FlowNotFound = CancelScheduledTasksResult &
     GetFlowResult & {
         __typename?: "FlowNotFound";
@@ -895,8 +906,16 @@ export type FlowStartCondition = FlowStartConditionBatching | FlowStartCondition
 
 export type FlowStartConditionBatching = {
     __typename?: "FlowStartConditionBatching";
-    thresholdNewRecords: Scalars["Int"];
+    accumulatedRecordsCount: Scalars["Int"];
+    activeBatchingRule: FlowConfigurationBatching;
+    batchingDeadline: Scalars["DateTime"];
+    watermarkModified: Scalars["Boolean"];
 };
+
+export enum FlowStartConditionKind {
+    Batching = "BATCHING",
+    Throttling = "THROTTLING",
+}
 
 export type FlowStartConditionThrottling = {
     __typename?: "FlowStartConditionThrottling";
@@ -2114,7 +2133,17 @@ export type FlowSummaryDataFragment = {
         finishedAt?: string | null;
     };
     startCondition?:
-        | { __typename: "FlowStartConditionBatching"; thresholdNewRecords: number }
+        | {
+              __typename: "FlowStartConditionBatching";
+              batchingDeadline: string;
+              accumulatedRecordsCount: number;
+              watermarkModified: boolean;
+              activeBatchingRule: {
+                  __typename?: "FlowConfigurationBatching";
+                  minRecordsToAwait: number;
+                  maxBatchingInterval: { __typename?: "TimeDelta" } & TimeDeltaDataFragment;
+              };
+          }
         | { __typename: "FlowStartConditionThrottling"; intervalSec: number }
         | null;
 };
@@ -2780,8 +2809,7 @@ export type DatasetFlowBatchingMutationVariables = Exact<{
     datasetId: Scalars["DatasetID"];
     datasetFlowType: DatasetFlowType;
     paused: Scalars["Boolean"];
-    throttlingPeriod?: InputMaybe<TimeDeltaInput>;
-    minimalDataBatch?: InputMaybe<Scalars["Int"]>;
+    batching: BatchingConditionInput;
 }>;
 
 export type DatasetFlowBatchingMutation = {
@@ -2801,6 +2829,7 @@ export type DatasetFlowBatchingMutation = {
                               expectedDatasetKind: DatasetKind;
                               actualDatasetKind: DatasetKind;
                           }
+                        | { __typename: "FlowInvalidBatchingConfig"; message: string; reason: string }
                         | {
                               __typename: "SetFlowConfigSuccess";
                               message: string;
@@ -2808,8 +2837,8 @@ export type DatasetFlowBatchingMutation = {
                                   __typename?: "FlowConfiguration";
                                   batching?: {
                                       __typename?: "FlowConfigurationBatching";
-                                      minimalDataBatch?: number | null;
-                                      throttlingPeriod?: ({ __typename?: "TimeDelta" } & TimeDeltaDataFragment) | null;
+                                      minRecordsToAwait: number;
+                                      maxBatchingInterval: { __typename?: "TimeDelta" } & TimeDeltaDataFragment;
                                   } | null;
                               };
                           };
@@ -2844,8 +2873,8 @@ export type GetDatasetFlowConfigsQuery = {
                                   | null;
                               batching?: {
                                   __typename?: "FlowConfigurationBatching";
-                                  minimalDataBatch?: number | null;
-                                  throttlingPeriod?: ({ __typename?: "TimeDelta" } & TimeDeltaDataFragment) | null;
+                                  minRecordsToAwait: number;
+                                  maxBatchingInterval: { __typename?: "TimeDelta" } & TimeDeltaDataFragment;
                               } | null;
                           } | null;
                       };
@@ -2879,6 +2908,7 @@ export type DatasetFlowScheduleMutation = {
                               expectedDatasetKind: DatasetKind;
                               actualDatasetKind: DatasetKind;
                           }
+                        | { __typename: "FlowInvalidBatchingConfig" }
                         | {
                               __typename: "SetFlowConfigSuccess";
                               message: string;
@@ -2944,6 +2974,12 @@ export const AccountFragmentDoc = gql`
         isAdmin
     }
 `;
+export const TimeDeltaDataFragmentDoc = gql`
+    fragment TimeDeltaData on TimeDelta {
+        every
+        unit
+    }
+`;
 export const FlowSummaryDataFragmentDoc = gql`
     fragment FlowSummaryData on Flow {
         description {
@@ -2996,11 +3032,20 @@ export const FlowSummaryDataFragmentDoc = gql`
                 intervalSec
             }
             ... on FlowStartConditionBatching {
-                thresholdNewRecords
+                activeBatchingRule {
+                    minRecordsToAwait
+                    maxBatchingInterval {
+                        ...TimeDeltaData
+                    }
+                }
+                batchingDeadline
+                accumulatedRecordsCount
+                watermarkModified
             }
         }
     }
     ${AccountFragmentDoc}
+    ${TimeDeltaDataFragmentDoc}
 `;
 export const DatasetPageInfoFragmentDoc = gql`
     fragment DatasetPageInfo on PageBasedInfo {
@@ -3832,12 +3877,6 @@ export const DatasetSearchOverviewFragmentDoc = gql`
     ${DatasetBasicsFragmentDoc}
     ${DatasetCurrentInfoFragmentDoc}
     ${LicenseFragmentDoc}
-`;
-export const TimeDeltaDataFragmentDoc = gql`
-    fragment TimeDeltaData on TimeDelta {
-        every
-        unit
-    }
 `;
 export const AccountByNameDocument = gql`
     query accountByName($accountName: AccountName!) {
@@ -4678,28 +4717,22 @@ export const DatasetFlowBatchingDocument = gql`
         $datasetId: DatasetID!
         $datasetFlowType: DatasetFlowType!
         $paused: Boolean!
-        $throttlingPeriod: TimeDeltaInput
-        $minimalDataBatch: Int
+        $batching: BatchingConditionInput!
     ) {
         datasets {
             byId(datasetId: $datasetId) {
                 flows {
                     configs {
-                        setConfigBatching(
-                            datasetFlowType: $datasetFlowType
-                            paused: $paused
-                            throttlingPeriod: $throttlingPeriod
-                            minimalDataBatch: $minimalDataBatch
-                        ) {
+                        setConfigBatching(datasetFlowType: $datasetFlowType, paused: $paused, batching: $batching) {
                             __typename
                             ... on SetFlowConfigSuccess {
                                 message
                                 config {
                                     batching {
-                                        throttlingPeriod {
+                                        maxBatchingInterval {
                                             ...TimeDeltaData
                                         }
-                                        minimalDataBatch
+                                        minRecordsToAwait
                                     }
                                 }
                             }
@@ -4707,6 +4740,10 @@ export const DatasetFlowBatchingDocument = gql`
                                 message
                                 expectedDatasetKind
                                 actualDatasetKind
+                            }
+                            ... on FlowInvalidBatchingConfig {
+                                message
+                                reason
                             }
                         }
                     }
@@ -4749,10 +4786,10 @@ export const GetDatasetFlowConfigsDocument = gql`
                                 }
                             }
                             batching {
-                                throttlingPeriod {
+                                maxBatchingInterval {
                                     ...TimeDeltaData
                                 }
-                                minimalDataBatch
+                                minRecordsToAwait
                             }
                         }
                     }