+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduced dataset scheduling settings page to manage automatically launched ingest/transform flows
 - Added new flows tab 
 - Added refresh now button on the overveiw tab to start the flow
+- Added support for new batching configuration
 ### Fixed
 - Fixed bug with authentication (Login vs CLI)
 - Settings tab is hidden after login